Ether Swap Sample to not use payable crosschain calls (#7)
Crosschain transactions (originating or subordinate transactions) can not be payable functions. This code changes the atomic swap ether sample to not have a payable function for the exchange. Instead, there is a deposit, withdraw, and get balance function that can be called before / after the atomic swap of ether occurs in the exchange function.

The libraries have been updated as WEB3J was iterated to not generate crosschain wrapper code for payable deployLockable and transaction functions.